Test rendering like button and count
We test that we're rendering the like button and count in the
PostComponent. Because the like button doesn't have a clear identifiable
string, we use a data-role to represent the element for our test.

We also modify the PostComponent's public interface by passing in a
`@current_user`. That causes our TimelineLiveTest to break in an
expected manner.

This breakage shows how our tests are independently responsible for
their own behaviors, while at the same time having that layer of
integration that ensures the PostComponents are being rendered correctly
from TimelineLive.