Change implementation to live patch
What changed?
============

We change the implementation of showing a highlighted post from using a
`phx-click` attribute to using `live_patch`.

In doing so, we show that we can keep the tests from the perspective of
the user without having to use `assert_patch`.

There are instances when we might want to use `assert_patch`, but even
in the cases where we want to make a URL a first-class concern, we can
write a separate test that makes that explicit.