Test displaying new posts
We test-drive the ability to click on the "Show X posts" notice. In the
test, we broadcast the post-created event (as though a different user
has created a post), and we interact with the UI by clicking on the
show-posts notice and expecting the new posts to be visible.

In the implementation, we now have to keep two lists in memory. The
`posts` list and the `new_posts` list. When we handle the
"show-new-posts" event, we combine the two lists and reset the
`new_posts` back to be empty.