Testing liking posts
We test that a user can like a post. We test it all through the
`TimelineLiveTest` even though we start seeing interactions between the
`PostComponent` and the `TimelineLive`.

For now, the interactions are minimal. The `PostComponent` continues to
be a static component, and though the `phx-click` attribute is there,
all the behavior is really handled by the `TimelineLive`.