Refactor source of truth to live components
What changed?
============

We refactor who is the source of truth for a `post`. Prior to this
commit, `TimelineLive` fetched all `posts` and passed each post to
`PostComponent` and `ShowPostComponent`.

This commit refactors the source of truth into the live components. So
the `TimelineLive` only fetches `ids` now, and it lets the components
fetch their own state. We use the `preload/1` function to avoid
performing N+1 queries when getting the posts.

Finally, since the `PostComponent` is its own source of truth, we remove
the sending of the message to the parent live view when a "like"
happens, and instead we update the `post` assign with the updated
version.