Broadcast like from PostComponent
We add a test to ensure we're broadcasting the updated post from the
`PostComponent`. We continue using the `Timeline` as the source of truth
for the PubSub topic and message format.

This is something of an intermediate step since we'll want to push the
logic into broadcasting logic into the `Timeline` itself.