Move broadcasting of like to Timeline.like_post!/2
We move the broadcasting of the like from the `PostComponent` and into
the `Timeline.like_post!/2` function. That is a better place for the
broadcast, since we always want to do that when a post is liked.

Since we move the broadcasting logic into the `Timeline`, we also move
the test to `TimelineTest`. That way, we keep the test as close to the
module that implements that behavior.